fix Infoscreen, credit
