+++ JamCrawler/src/components/Instruction/InfoScreen.css
